requirements.txt
Django==5.1.3
djangorestframework==3.15.2
mysqlclient==2.2.5
django-mysql==4.15.0
django-cors-headers==4.6.0
pika==1.3.2
requests==2.32.3
drf-spectacular==0.28.0
python-dotenv==1.0.1
grpcio==1.68.0
grpcio-tools==1.68.0
gunicorn==23.0.0
django-dotenv==1.4.2
psycopg2-binary==2.9.10
django-storages==1.14.4
boto3==1.35.76
pytest==8.3.4